requirements.txt
certifi==2024.7.4
charset-normalizer==3.3.2
einops==0.8.0
filelock==3.15.4
fsspec==2024.6.1
huggingface-hub==0.24.6
idna==3.8
Jinja2==3.1.4
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
packaging==24.1
pillow==10.4.0
PyYAML==6.0.2
regex==2024.7.24
requests==2.32.3
safetensors==0.4.4
sympy==1.13.2
timm==1.0.9
tokenizers==0.19.1
torch==2.2.2
torchvision==0.17.2
tqdm==4.66.5
transformers==4.44.2
typing_extensions==4.12.2
urllib3==2.2.2
